ENH: add windows compile related changes
    1.add /W4 instead of /Wall to avoid like
    "unused inline func...remove" silly warnings
    2.Notice different CXX_FLAGS for VS2022
    3.fix some warnings in VIEO_SLAM
    4.add windows macro for g2o(refer
    to newest g2o codes) to let compile
    on windows pass
    5.add access && mkdir for windows
    6.fix our FindEigen3.cmake for windows
    0.TODO: set affinity linux 1

+++ include/ORBextractor.h
@@ -41,7 +41,7 @@ class ORBextractor {
 
   int inline GetLevels() { return nlevels; }
 
-  float inline GetScaleFactor() { return scaleFactor; }
+  float inline GetScaleFactor() { return (float)scaleFactor; }
 
   std::vector<float> inline GetScaleFactors() { return mvScaleFactor; }
